+++ NEWS.rst
@@ -7,3 +7,6 @@ every change, see the Git log.
 Latest
 ------
 * Major: Added level member to metrics class
+* Minor: Added read_storage() to metrics class
+* Major: Separated the metric class from metrics
+